+++ crates/cli/src/main.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Google LLC
+// Copyright 2023 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,25 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//{ ANCHOR: all
-//{ ANCHOR: nostd
-#![no_std]
-//} ANCHOR_END: nostd
-
-//{ ANCHOR: alloc
-extern crate alloc;
-//} ANCHOR_END: alloc
-
-//{ ANCHOR: prelude
-use wasefire::*;
-//} ANCHOR_END: prelude
-
-//{ ANCHOR: nomangle
-#[no_mangle]
-//} ANCHOR_END: nomangle
-//{ ANCHOR: externc
-pub extern "C" fn main() {
-//} ANCHOR_END: externc
-    println!("hello world");
+fn main() {
+    println!("The Wasefire CLI is not yet implemented.");
+    std::process::exit(1);
 }
-//} ANCHOR_END: all