+++ crates/runner-nordic/src/tasks/mod.rs
@@ -46,6 +46,10 @@ impl board::Api for Board {
         }
     }
 
+    fn breakpoint(&mut self) {
+        cortex_m::asm::bkpt();
+    }
+
     fn take_storage(&mut self) -> Option<Self::Storage> {
         critical_section::with(|cs| self.0.borrow_ref_mut(cs).storage.take())
     }