+++ examples/rust/button/src/lib.rs
@@ -17,13 +17,9 @@
 //! The applet prints the button state to the debug output on any button event.
 
 #![no_std]
+wasefire::applet!();
 
-extern crate alloc;
-
-use wasefire::*;
-
-#[no_mangle]
-pub extern "C" fn main() {
+fn main() {
     // Make sure there is at least one button.
     let count = button::count();
     assert!(count > 0, "Board has no buttons.");
@@ -31,7 +27,7 @@ pub extern "C" fn main() {
     // For each button on the board.
     for index in 0 .. count {
         // We define a button handler printing the new state.
-        let handler = move |state| println!("Button {index} has been {state:?}.");
+        let handler = move |state| debug!("Button {index} has been {state:?}.");
 
         // We start listening for state changes with the handler.
         let listener = button::Listener::new(index, handler);