+++ examples/rust/led/src/lib.rs
@@ -21,16 +21,12 @@
 //! to the first LED after the last LED).
 
 #![no_std]
-
-extern crate alloc;
+wasefire::applet!();
 
 use alloc::boxed::Box;
 use core::cell::Cell;
 
-use wasefire::*;
-
-#[no_mangle]
-pub extern "C" fn main() {
+fn main() {
     // Make sure there is at least one button.
     let num_buttons = button::count();
     assert!(num_buttons > 0, "Board has no buttons.");