+++ crates/api-desc/CHANGELOG.md
@@ -1,7 +1,23 @@
 # Changelog
 
+## 0.1.2-git
+
 ### Minor
 
+- Add scheduling::breakpoint
+
+### Patch
+
+- Update dependencies
+- Test that link names are unique
+
+## 0.1.1
+
+### Minor
+
+- Support wasefire-applet-api test feature
 - Add copyright header on generated files
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->