+++ crates/api-macro/CHANGELOG.md
@@ -1,3 +1,17 @@
 # Changelog
 
+## 0.1.2-git
+
+### Patch
+
+- Update dependencies
+
+## 0.1.1
+
+### Minor
+
+- Update wasefire-applet-api-desc version
+
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #112 -->