+++ crates/api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.1.3-git
+
+### Patch
+
+- Update Cargo.lock file (it is packaged due to
+  https://github.com/rust-lang/cargo/issues/11557 although examples are not)
+
+## 0.1.2
+
+### Minor
+
+- Update wasefire-applet-api-macro version
+- Add test feature panicking on applet API calls
+
 ## 0.1.1
 
 ### Patch
@@ -7,3 +21,5 @@
 - Build `docs.rs` with `wasm` feature on `thumbv7em-none-eabi` target
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #112 -->