+++ crates/board/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 0.1.3-git
+
+### Minor
+
+- Add breakpoint in Api
+
+### Patch
+
+- Update dependencies
+
+## 0.1.2
+
+### Patch
+
+- Update wasefire-logger version
+
 ## 0.1.1
 
 ### Patch
@@ -7,3 +23,5 @@
 - Build `docs.rs` with `std` feature
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #112 -->