+++ crates/interpreter/CHANGELOG.md
@@ -1,3 +1,22 @@
 # Changelog
 
+## 0.1.2-git
+
+### Minor
+
+- Use `atomic-polyfill` to support RV32IMC
+
+### Patch
+
+- Fix use of deprecated function
+- Update dependencies
+
+## 0.1.1
+
+### Patch
+
+- Panic if `start` calls a host function
+
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->