+++ crates/logger/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.2-git
+
+### Patch
+
+- Update dependencies
+
+## 0.1.1
+
 ### Minor
 
 - Add `Display2Format` and `panic`
@@ -9,3 +17,5 @@
 - Fix copyright header position
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->