+++ crates/prelude/CHANGELOG.md
@@ -1,3 +1,31 @@
 # Changelog
 
+## 0.1.4-git
+
+### Minor
+
+- Add scheduling::breakpoint
+
+## 0.1.3
+
+### Patch
+
+- Update documentation with 0.1.2 change
+
+## 0.1.2
+
+### Minor
+
+- Add `debug!` macro as a replacement for `println!`
+- Deprecate `println!` macro in favor of `debug!`
+
+## 0.1.1
+
+### Minor
+
+- Add test feature
+- Add applet!() macro
+
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->