+++ crates/scheduler/CHANGELOG.md
@@ -1,7 +1,21 @@
 # Changelog
 
+## 0.1.2-git
+
+### Minor
+
+- Support scheduling::breakpoint
+
+### Patch
+
+- Update dependencies
+
+## 0.1.1
+
 ### Patch
 
 - Use `logger::panic!` instead of `core::panic!`
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #112 -->